A revolutionary update (#1)

+++ scripts/dev-server.js
@@ -37,11 +37,19 @@ async function startFlask() {
     // single instance lock
     return
   }
+  console.log(Chalk.cyanBright(`[python] `) + 'Loading...')
 
   const jmPath = Path.resolve(jianmuPath, 'jm.py')
+  const srcPath = Path.resolve(projectPath, 'src')
+  const PYTHONPATH = process.env.PYTHONPATH
+    ? `${srcPath}:${projectPath}:${process.env.PYTHONPATH}`
+    : projectPath
   flaskProcess = ChildProcess.spawn(pythonPath, [jmPath], {
-    cwd: projectPath,
-    env: process.env
+    cwd: srcPath,
+    env: {
+      ...process.env,
+      PYTHONPATH
+    }
   })
   flaskProcessLocker = false
 
@@ -89,7 +97,33 @@ async function startFlask() {
     }
   })
 
-  flaskProcess.on('exit', () => stop())
+  flaskProcess.on('exit', () => {
+    if (isDev && electronProcess && electronProcess.exitCode === null) {
+      // 如果 Electron 还没退出且处于开发模式，就重启 Flask 服务。
+      // Clear outputs in the console.
+      // process.stdout.write('\x1Bc')
+      // console.log(
+      //   Chalk.cyanBright(`[python] `) +
+      //     'Reloading failed, try to reload again after 1 seconds...'
+      // )
+      // setTimeout(() => {
+      //   restartFlask()
+      // }, 1000)
+      /**
+       * NOTE: 根据反馈，这个自动重启 Flask 服务的功能有两个问题：
+       * ① 会导致报错信息丢失，因为重启之后，之前的报错信息就没了。
+       * ② 会让用户在一些情况下感到困惑，比如检测到目录下文件变动后，Flask服务重启了，然后就和未重启的 Electron 断开链接了
+       * 因此目前先注释掉这个功能，后续再考虑是否要加回来。
+       */
+      console.log(
+        Chalk.cyanBright(`[python] `) +
+          'The python process has exited unexpectedly, so the total program will exit now.'
+      )
+      stop()
+    } else {
+      stop()
+    }
+  })
 }
 
 function restartFlask() {
@@ -123,6 +157,8 @@ async function startElectron(copyStaticFiles = true) {
     return
   }
 
+  console.log(Chalk.cyanBright(`[ui] `) + 'Starting...')
+
   const args = [
     Path.resolve(__dirname, '..', '.jianmu', 'electron', 'main.js'),
     rendererPort
@@ -142,6 +178,9 @@ async function startElectron(copyStaticFiles = true) {
       if (!msg.trim()) {
         continue
       }
+      if (msg.match(/Electron\[\d+:\d+\]\s/)) {
+        continue
+      }
       process.stdout.write(prefix + msg)
     }
   })
@@ -154,6 +193,9 @@ async function startElectron(copyStaticFiles = true) {
         continue
       }
       if (!msg.trim()) {
+        continue
+      }
+      if (msg.match(/Electron\[\d+:\d+\]\s/)) {
         continue
       }
       process.stderr.write(prefix + msg)
@@ -172,21 +214,23 @@ function restartElectron() {
 
   if (!electronProcessLocker) {
     electronProcessLocker = true
+    // Clear outputs in the console.
+    process.stdout.write('\x1Bc')
     startElectron(false)
   }
 }
 
 function stop() {
   console.log(Chalk.redBright('Stop Jianmu Development Server...'))
-  if (isDev) {
+  if (isDev && viteServer) {
     viteServer.close()
   }
-  if (electronProcess.exitCode === null) {
+  if (electronProcess && electronProcess.exitCode === null) {
     electronProcess.removeAllListeners('exit')
     electronProcess.kill()
     electronProcess = null
   }
-  if (flaskProcess.exitCode === null) {
+  if (flaskProcess && flaskProcess.exitCode === null) {
     flaskProcess.removeAllListeners('exit')
     flaskProcess.kill()
     flaskProcess = null
@@ -214,18 +258,26 @@ async function start(_pythonPath, _jianmuPath, _projectPath, _isDev) {
     process.env.NODE_ENV = 'production'
   }
 
+  // Clear outputs in the console.
+  process.stdout.write('\x1Bc')
   printBanner()
 
   if (isDev) {
-    console.log(Chalk.greenBright('Start Jianmu Server in development mode.\n'))
+    console.log(
+      Chalk.greenBright('Start Jianmu Application in development mode.\n')
+    )
   } else {
-    console.log(Chalk.greenBright('Start Jianmu Server in production mode.'))
+    console.log(
+      Chalk.greenBright('Start Jianmu Application in production mode.')
+    )
     console.log(
       Chalk.yellowBright('[WARNING] ') +
-        'In production mode, the server will not reload on file changes.\n'
+        'In production mode, the Jianmu Application will not reload on file changes.\n'
     )
   }
-  console.log(Chalk.yellowBright('You can press Ctrl+C to stop the server.\n'))
+  console.log(
+    Chalk.yellowBright('You can press Ctrl+C to stop the Jianmu Application.\n')
+  )
 
   // 删除临时文件目录 .jianmu
   await emptyTempDir()
@@ -242,21 +294,27 @@ async function start(_pythonPath, _jianmuPath, _projectPath, _isDev) {
   // 启动 Flask 服务
   startFlask(pythonPath, jianmuPath, projectPath)
 
+  /**
+   * NOTE: 根据反馈，这个自动重启 Flask 服务的功能有两个问题：
+   * ① 会导致报错信息丢失，因为重启之后，之前的报错信息就没了。
+   * ② 会让用户在一些情况下感到困惑，比如检测到目录下文件变动后，Flask服务重启了，然后就和未重启的 Electron 断开链接了
+   * 因此目前先注释掉这个功能，后续再考虑是否要加回来。
+   */
   // 如果是 development 模式，侦听文件变化并重启 Flask 服务
-  if (isDev) {
-    const pythonSrcPath = Path.resolve(projectPath, 'src')
-    Chokidar.watch(pythonSrcPath, {
-      cwd: pythonSrcPath
-    }).on('change', (path) => {
-      if (path.endsWith('.pyc') || path.endsWith('.tmp')) {
-        return
-      }
-      console.log(
-        Chalk.cyanBright(`[python] `) + `Change in ${path}. reloading...`
-      )
-      restartFlask()
-    })
-  }
+  // if (isDev) {
+  //   const pythonSrcPath = Path.resolve(projectPath, 'src')
+  //   Chokidar.watch(pythonSrcPath, {
+  //     cwd: pythonSrcPath
+  //   }).on('change', (path) => {
+  //     if (path.endsWith('.pyc') || path.endsWith('.tmp')) {
+  //       return
+  //     }
+  //     console.log(
+  //       Chalk.cyanBright(`[python] `) + `Change in ${path}. reloading...`
+  //     )
+  //     restartFlask()
+  //   })
+  // }
 
   // 编译 Electron Main 进程文件，并启动 Electron
   startElectron()
